json issue_2

+++ backend/app.py
@@ -321,17 +321,14 @@ def run_optimization():
     total_cost_all = df_full['Total_Cost'].sum()
     channel_summary['% of Total'] = (channel_summary['Total_Cost'] / total_cost_all * 100).round(2)
 
-
     return jsonify({
         "success": True,
-        "total_cost": round(total_cost_all, 2),
+        "total_cost": float(round(total_cost_all, 2)),
         "total_rating": float(round(df_full['Total_Rating'].sum(), 2)),
         "cprp": float(round(total_cost_all / df_full['Total_Rating'].sum(), 2)) if df_full['Total_Rating'].sum() else None,
         "commercials_summary": commercials_summary,
-        #"channel_summary": channel_summary.to_dict(orient='records'),
         "channel_summary": json.loads(channel_summary.to_json(orient='records')),
         "df_result": json.loads(df_full.to_json(orient='records'))
-
     })
